+++ cah/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = '2.0.6'
-__update_date__ = '2022-04-15_17:54:40'
+__version__ = '2.0.7'
+__update_date__ = '2022-04-15_19:41:22'
 from pathlib import Path
 ROOT_PATH = Path(__file__)